requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
numpy==1.26.2
psycopg2-binary==2.9.9
sqlalchemy-utils==0.41.1
